Change "Search strategy" to "Databases to search" [#3]
"Search strategy" was a misnomer. The `DEFINE` command takes a database
to search rather than a search strategy. The `MATCH` command, which is
not currently implemented here, is the one to use with a search
strategy.

This change makes that more clear. The `MATCH` functionality with search
strategy can be implemented later.

Also de-duplicate some UI strings.